requirements.txt
amqp==5.1.0
asgiref==3.5.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.23.0
azure-storage-blob==12.10.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==3.4.8
Django==4.0.3
django-compat==1.0.15
django-cors-headers==3.11.0
django-environ==0.8.1
django-jazzmin==2.5.0
django-storages==1.12.3
django4-background-tasks==1.2.7
djangorestframework==3.13.1
docx2txt==0.8
flake8==4.0.1
idna==3.3
isodate==0.6.1
jsons==1.6.1
kombu==5.2.4
lxml==4.8.0
mccabe==0.6.1
msal==1.10.0
msrest==0.6.21
numpy==1.22.3
oauthlib==3.2.0
pandas==1.4.1
Pillow==9.0.1
prompt-toolkit==3.0.28
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
python-dateutil==2.8.1
python-docx==0.8.11
python-dotenv==0.19.2
pytz==2021.3
PyYAML==5.4.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
typish==1.9.3
tzdata==2021.5
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
